simplifying logical flow of client.js by spliting out html generating functions into content.js
